+++ wilfred/core.py
@@ -14,7 +14,7 @@ import click
 from random import choice
 from string import ascii_lowercase, digits
 
-from wilfred.version import version
+from wilfred.version import version, commit_hash
 from wilfred.message_handler import warning
 
 
@@ -33,7 +33,16 @@ def check_for_new_releases(enable_emojis=True):
     Checks if a new version is available on GitHub
     """
 
-    r = requests.get("https://api.github.com/repos/wilfred-dev/wilfred/tags")
+    url = "https://api.github.com/repos/wilfred-dev/wilfred/tags"
+    key = "name"
+    version_type = "version"
+
+    if version == "0.0.0.dev0":
+        url = "https://api.github.com/repos/wilfred-dev/wilfred/commits"
+        key = "sha"
+        version_type = "commit"
+
+    r = requests.get(url)
 
     if r.status_code != requests.codes.ok:
         warning("unable to retrieve latest version")
@@ -41,7 +50,7 @@ def check_for_new_releases(enable_emojis=True):
         return
 
     try:
-        latest_release = r.json()[0]["name"]
+        latest_release = r.json()[0][key]
     except Exception:
         warning("unable to parse release data")
 
@@ -52,7 +61,7 @@ def check_for_new_releases(enable_emojis=True):
             "".join(
                 (
                     f"{'🎉 ' if enable_emojis else click.style('! ', fg='green')}",
-                    f"A new version of Wilfred is available! {latest_release}",
+                    f"A new {version_type} of Wilfred is available! {latest_release}",
                 )
             )
         )